requirements.txt
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.4
